Rename modules directories.
