+++ data/src/main/java/com/saulgarcia/data/mapper/MealsCategoryMapper.kt
@@ -1,0 +1,6 @@
+package com.saulgarcia.data.mapper
+
+import com.saulgarcia.domain.entity.MealsCategory
+import com.saulgarcia.data.entity.MealsCategoryResponse
+
+fun MealsCategoryResponse.toMealsCategory() = MealsCategory(id = id)
