+++ data/src/main/java/com/saulgarcia/data/repository/MealsCategoryRepositoryImpl.kt
@@ -1,10 +1,10 @@
-package com.saulgarcia.dataq.repository
+package com.saulgarcia.data.repository
 
-import com.saulgarcia.domainq.entity.action.Either
-import com.saulgarcia.domainq.entity.action.error.ErrorEntity
-import com.saulgarcia.dataq.mapper.toMealsCategory
-import com.saulgarcia.dataq.remote.TheMealApiClient
-import com.saulgarcia.domainq.repository.MealsCategoryRepository
+import com.saulgarcia.domain.entity.action.Either
+import com.saulgarcia.domain.entity.action.error.ErrorEntity
+import com.saulgarcia.data.mapper.toMealsCategory
+import com.saulgarcia.data.remote.TheMealApiClient
+import com.saulgarcia.domain.repository.MealsCategoryRepository
 import javax.inject.Inject
 
 class MealsCategoryRepositoryImpl @Inject constructor (private val client: TheMealApiClient):
