--- data/src/main/java/com/saulgarcia/dataq/mapper/MealsCategoryMapper.kt
@@ -1,6 +1,0 @@
-package com.saulgarcia.dataq.mapper
-
-import com.saulgarcia.domainq.entity.MealsCategory
-import com.saulgarcia.dataq.entity.MealsCategoryResponse
-
-fun MealsCategoryResponse.toMealsCategory() = MealsCategory(id = id)
