+++ data/src/test/java/com/saulgarcia/data/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.saulgarcia.dataq
+package com.saulgarcia.data
 
 import org.junit.Test
