+++ domain/src/main/java/com/saulgarcia/domain/entity/MealsCategory.kt
@@ -1,4 +1,4 @@
-package com.saulgarcia.domainq.entity
+package com.saulgarcia.domain.entity
 
 data class MealsCategory(
     val id: Int? = null
