+++ domain/src/main/java/com/saulgarcia/domain/repository/MealsCategoryRepository.kt
@@ -1,0 +1,7 @@
+package com.saulgarcia.domain.repository
+
+import com.saulgarcia.domain.entity.action.Either
+
+interface MealsCategoryRepository {
+    fun getMealsCategories(): Either
+}
