+++ domain/src/main/java/com/saulgarcia/domain/useCase/GetCategoriesUseCase.kt
@@ -1,7 +1,7 @@
-package com.saulgarcia.domainq.useCase
+package com.saulgarcia.domain.useCase
 
-import com.saulgarcia.domainq.entity.action.Either
-import com.saulgarcia.domainq.repository.MealsCategoryRepository
+import com.saulgarcia.domain.entity.action.Either
+import com.saulgarcia.domain.repository.MealsCategoryRepository
 import javax.inject.Inject
 
 class GetCategoriesUseCase @Inject constructor(private val mealsRepository: MealsCategoryRepository){
