--- domain/src/main/java/com/saulgarcia/domainq/repository/MealsCategoryRepository.kt
@@ -1,7 +1,0 @@
-package com.saulgarcia.domainq.repository
-
-import com.saulgarcia.domainq.entity.action.Either
-
-interface MealsCategoryRepository {
-    fun getMealsCategories(): Either
-}
